refactor: imports reflecting recent project refactorings
