+++ web/apps/nextjs__raw_krabby_chat_api_integrations/app/(routes)/(chat)/components/MobileBottomNav/index.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import { ChatsIcon, CallsIcon, ContactsIcon, SettingsIcon } from '../../../../Icons';
+import { ChatsIcon, CallsIcon, ContactsIcon, SettingsIcon } from '../../../../icons';
 
 function MobileBottomNav() {
   return (
