+++ web/apps/nextjs__raw_krabby_chat_api_integrations/app/(routes)/(chat)/components/SideBar/index.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import UserAccountMenu from '../UserAccountMenu';
-import { ChatsIcon, ContactsIcon, CallsIcon, ArchiveIcon, SettingsIcon } from '../../../../Icons';
+import { ChatsIcon, ContactsIcon, CallsIcon, ArchiveIcon, SettingsIcon } from '../../../../icons';
 
 function Sidebar() {
   return (
